Fewer extra digits in arbitrary reals
It's a bit of a pain, but basically the arbitrary instance could
generate for precision p:

2, 1, 3. Where 2 <= 1 and 2<=3 (because they compare equal at the low
precision p) and hence 1 <= 3 (which is detected and fails the test).
Making the number of bits from the arbitrary instance match p sorts this
out by disallowing that extra precision.

Fixes https://github.com/expipiplus1/exact-real/issues/35